"Added rendering frame limiter and updated rendering to occur at a rate of 1000ms"
This commit adds a new variable `time_since_render` to track the time since the last frame was rendered. The rendering function `render()` is now called less frequently by wrapping it in a conditional that checks if the render rate (1000ms) has been exceeded since the last frame was rendered. If the render rate has been exceeded, the frame is rendered and the `time_since_render` variable is reset to 0.