"Added showGlobalRates function to display currentRates.js globalRates object"
The diff includes changes to the scripts/json/currentRates.js file. Specifically, it adds a new function called showGlobalRates which will iterate over the entries in the globalRates object and print them to the console. Additionally, the window object now has a reference to this new function, allowing it to be called from other parts of the code. The number of entries in the globalRates object is also logged to the console. The rest of the code remains unchanged.